Internalizes the new Anchor logic and removes it from the IContentService interface into extensions.
