+++ src/Umbraco.Core/Services/IContentService.cs
@@ -526,8 +526,6 @@ namespace Umbraco.Core.Services
         OperationResult Rollback(int id, int versionId, string culture = "*", int userId = Constants.Security.SuperUserId);
 
         #endregion
-        
-        IEnumerable<string> GetAnchorValuesFromRTEs(int id, string culture = "*");
-        IEnumerable<string> GetAnchorValuesFromRTEContent(string rteContent, string culture = "*");  
+      
     }
 }
